Reduce run-configuration manipulation after launching from product page
Noble (but failed) attempt at declaring all plugins in a
separate feature, so as to include this as a dependency to
the product.

Got stuck at the javax.servlet where I needed to specify an older
version out of two available versions. And the feature.xml does not
seem to support this...

To be continued, surely