Commit failed, but interesting attempt at adding control in top trim
Going to abandon this path, because i can't seem to manage to
control the height of the top trim bar. A solution would probably
be to create our own 'renderer' to replace the TrimBarLayout, but
for now I'm going to go a different path (fixed view in top).

Committing the work here, for easier retrieval afterwards in case
we make another attempt at it.