Show all times in UI using local TZ. Eventually should be configurable
There was a mix of UTC and local times. The reason being that the XYChart
used in OPIBuilder doesn't seem to allow any sort of timezone customization.

It only seems to accept a date format, which is not enough.

Current solution is to make the UI consistent by matching everything with
the default timezone (which is configurable through java itself). But
eventually we should make a user preference for this.