Fork some stuff as a preparation for #11
This removes the dependency of org.yamcs.studio.ui on utility.product.
We can't remove the plugin from the build though, due to it being
requested by another dependency. This needs to be researched further.

Next step would be to either hook the lifecyclemanager logic into
these classes, or to start taking over the startup class, which is
still an upstream dependency.