Revert tycho to 0.22.0 to fix OSX support
Apparently, and quite shockingly, Tycho 0.23.0+ no longer
supports building working OSX products that do not target
the Eclipse Mars project yet.

The unresolved issues is documented here:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=461606

Until either we move to Eclipse Mars, or Tycho fixes their
compatibility issues with OSX, we should not use the
latest version of Tycho.

A new branch was created for this reason to keep a
memento of where we were before this change.