Introduce ConnectionManager and start refactoring towards it
The YamcsPlugin was loaded with connection-related info. Ideally
we shouldn't even have a YamcsPlugin. This commits starts a big
refactoring of moving all the connection stuff in a (big)
ConnectionManager.

These refactorings are done in pair with the changes required
to use the new-and-improved system-wide connection preferences.

The build is broken, but I want to apply some other local branches
before moving on (Removing some hornetq stuff).