Update css deps to 4.2.x branch
This also changes the packaging due to upstream changes. Notably,
the diirt repository has been added and this does not read the
cs-studio-local property, hence we now use a different solution for
local p2 dependencies, which is to use a local composite p2 repository
that refers to the individual p2 repos of every project.