Command History: workaround a weird hang
When "Command History" is set to "Show key columns only" it seems
to hang when there are more than 8-10 lines in the Command History.

Reproducible on Ubuntu 20.04+Xorg+openbox+LXQt and Debian 10+Xorg+xfwm4+LXQt
with openjdk-8-jdk and with Sun Java 8 with CDMCS.

The proper fix would be to rebuild the Command History layout
with or without those columns, instead of just hiding them.